+++ examples/java-matter-controller/java/src/com/matter/controller/commands/pairing/PairAddressPaseCommand.kt
@@ -30,7 +30,7 @@ class PairAddressPaseCommand(controller: ChipDeviceController, credsIssue: Crede
     PairingModeType.ADDRESS_PASE_ONLY,
     PairingNetworkType.NONE
   ) {
-  override fun onPairingComplete(errorCode: Int) {
+  override fun onPairingComplete(errorCode: Long) {
     logger.log(Level.INFO, "onPairingComplete with error code: $errorCode")
     if (errorCode == 0) {
       setSuccess()
